Remove `repyhelper` from `repyportability`
See SeattleTestbed/portability#26.

In RepyV1, `repyhelper` would go and create a Python-importable file
out of a Repy file by
* prepending a few necessary `import`s of the Repy API libs etc., see
https://github.com/SeattleTestbed/portability/blob/master/repyhelper.py#L239-L241,
* writing that and the original file contents into a temp file
("LIBNAME_repy.py") so that Python's `import` function would find it, and
* then calling `__import__` to actually pull the Repy lib into the
current context.

In RepyV2, only `repyportability` would use `repyhelper`, and only to
pull in `dylink` (in the `add_dy_support()` function).

My attempt here is to emulate what `dylink` does when it is instructed
to pull in a library:
* Read in the code
* Create a RepyV2 VirtualNamespace around it
* Supply the Repy API to that namespace when evaluating it
* Save the evaluated namespace (i.e. a dict containing the function names
and references, besides other stuff internal to the library)
* Make the newly-defined `dylink` function calls available to the
caller's namespace.

The last point is trivial for `dylink` as it defines a function to do
precisely that: `init_dylink`.

From a bit of interactive testing, it seems that `dy_import_module` and
`dy_import_module_symbols` do the right thing, and we managed to get rid
of a good amount of overrides/workarounds that were required before.
However, I need to think more about thoroughly testing this, because
we rely on this functionality in critical places such as the Clearinghouse
code....